+++ CHANGELOG.md
@@ -1,10 +1,22 @@
 # Changelog
 
+## [0.0.5] - 2026-08-14
+
+### Added
+- New options for `reqtrace scan`: `--all` and `--requirements`
+	- `--all` shows packages without requirements
+	- `--requirements` prints paths to found REQUIREMENTS.md files only
+- Parse requirements now supports multiple files through stdin
+
+### Changed
+- Abstracted scan and parse logic to helper functions
+- Scan no longer shows packages without requirements by default (see: `-all`)
+
 ## [0.0.4] - 2026-08-14
 
 ### Added
-- Added parse to reqtrace cli (usage: reqtrace parse)
-- Implemented parsing for requirements file (usage: reqtrace parse requirements FILE)
+- Added parse to reqtrace cli (usage: `reqtrace parse`)
+- Implemented parsing for requirements file (usage: `reqtrace parse requirements FILE`)
 
 ## [0.0.3] - 2026-08-09
 
@@ -13,7 +25,7 @@
 - Wrote logic for package discovery in discover.py
 
 ### Changed
-- Updated cli scan command to show results of package discovery
+- Updated `cli scan` command to show results of package discovery
 
 ## [0.0.2] - 2026-08-09